pyproject.toml
[project]
name = "nongo"
version = "0.1.0"
description = "Network + funding mapping for Norwegian political/diplomatic ties"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "openpyxl>=3.1.5",
  "psycopg[binary]>=3.2.0",
  "neo4j>=5.28.0",
  "python-dotenv>=1.0.1",
  "requests>=2.32.0",
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "jinja2>=3.1.4",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.9.0",
  "pytest>=8.3.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
